+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changes
 - send normal messages with higher priority than MDNs #3243
+- make Scheduler stateless #3302
 
 
 ## 1.80.0